rename factories files
in prep for creating a separate function to create tasks from json
